pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = ""
authors = ["nervevau2 <eugenshow83@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7 <=3.10"
djangorestframework = "==3.12.4"
Django = "==3.2.13 "
drf-yasg = "==1.20.0"
jsonschema = "==3.2.0"
django-cors-headers = "==3.12.0"
django-filter = "==21.1"
swagger_spec_validator = "==2.7.4"
python-dotenv = "==0.20.0"
django-environ = "==0.8.1"
postgres  = "==4.0"
django-rest-swagger = "==2.2.0"
simplejwt = "==2.0.1"
djangorestframework-simplejwt = "==4.8.0"
drf-util =  "==1.0.14"
PyJWT = "==2.4.0"
Faker = "==13.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
